Migrate fjs/effects/memory to .f.mjs, split type-level API into types.ts
Leaf module. Updates 13 dependents across cas, mcp, protocol/mcp, and
effects/node to import runtime values from .f.mjs and types from
types.ts separately, and fixes two markdown link hrefs.
