+++ fjs/cas/evo/module.f.ts
@@ -7,7 +7,7 @@
  * are the current heads of subject X" from that store means walking every
  * stored revision and reversing the parent links — too expensive to redo on
  * every query. This module scans the store once into an in-memory `Cache`
- * ([`fjs/effects/memory`](../../effects/memory/module.f.ts)) mapping subject →
+ * ([`fjs/effects/memory`](../../effects/memory/module.f.mjs)) mapping subject →
  * head hashes, then keeps the cache current as new revisions are `add`ed
  * through it — or, for a revision stored some other way (e.g. the generic
  * `cas_add` MCP tool), via {@link syncRevision}.
@@ -45,7 +45,8 @@
 import { pure, foldStep } from '../../effects/module.f.mjs'
 import type { Effect, Operation } from '../../effects/types.ts'
 import { eff } from '../../effects/eff/module.f.ts'
-import { create, read, write, type Key, type MemOp } from '../../effects/memory/module.f.ts'
+import { create, read, write } from '../../effects/memory/module.f.mjs'
+import type { Key, MemOp } from '../../effects/memory/types.ts'
 import { collectRead, type Cas } from '../module.f.ts'
 import { cBase32ToVec, vecToCBase32 } from '../../basen/cbase32/module.f.mjs'
 import { fromVec } from '../../text/utf8/module.f.mjs'
