+++ fjs/effects/node/module.f.ts
@@ -15,7 +15,7 @@ import type { List as EffectList } from '../../types/list/types.ts'
 import { reverse } from '../../types/list/module.f.mjs'
 import type { Vec } from '../../types/bit_vec/types.ts'
 import { length } from '../../types/bit_vec/module.f.mjs'
-import type { MemOp } from '../memory/module.f.ts'
+import type { MemOp } from '../memory/types.ts'
 import type { Nominal } from '../../types/nominal/types.ts'
 import type { Result } from '../../types/result/types.ts'
 import { ok, error as resultError, mapOk } from '../../types/result/module.f.mjs'
