+++ fjs/effects/node/virtual/module.f.ts
@@ -10,7 +10,8 @@ import type { Vec } from '../../../types/bit_vec/types.ts'
 import { empty, length, maxLengthBytes, msb, vec } from '../../../types/bit_vec/module.f.mjs'
 import { error, ok } from '../../../types/result/module.f.mjs'
 import { run, type MemOperationMap, type RunInstance } from '../../mock/module.f.ts'
-import { asBase, asNominal, type Key } from '../../memory/module.f.ts'
+import { asBase, asNominal } from '../../memory/module.f.mjs'
+import type { Key } from '../../memory/types.ts'
 import type { Dirent, FileStat, IoResult, Module, NodeOp, NodeProgramOptions, SandboxResult } from '../module.f.ts'
 
 /**
